+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 None.
 
 ## Latest release: 0.8.0
+* [knife-windows #96](https://github.com/opscode/knife-windows/issues/96) Fix break from OS patch KB2918614
 * Update winrm-s dependency along with em-winrm and winrm dependencies
 * Return failure codes from knife winrm even when `returns` is not set
 * Support Windows negotiate authentication protocol when running knife on Windows